TM_supp/email_python/requirements.txt
# needed for later use with Cloud deployment ( I think... )
pandas
python-dotenv